backend/requirements.txt
Django>=4.2,<5.0
djangorestframework==3.16.0
django-cors-headers>=4.3.0

